apps/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
python-socketio==5.11.0
openai==1.9.0
langchain==0.1.1
whisper==1.1.10
pyaudio==0.2.14
pytest==7.4.4
pytest-asyncio==0.23.3
